Decrease the numOctaves to 3 for performance reasons (PR #23)
This commit decreases the numOctaves used in the feTurbulence
fractalNoise SVG filter from 10 to 3. @zpiman investigated this in
PR #23 and found it to have a significant impact on SVG render time.
While I found that reducing it all the way down to 1 caused a visual
impact on the generated noise, 3 seems to be a good compromise that
improves performance significantly while staying visually
indistinguishable from 10.